fix(runtime): march_task_await spin safety + cancel-token ABI + dedup fwd-decl
- march_task_await: add sched_yield() fallback so the spin doesn't burn
  CPU when called from outside a green-thread context (march_sched_yield
  is a no-op there)
- march_cancel_token_is_cancelled: change return type int → int64_t in
  both the .h declaration and .c definition so it matches the LLVM
  `declare i64` — was an ABI mismatch (harmless on x86-64 but wrong)
- Remove duplicate static forward declarations for mk_ok/mk_err_cstr
  at the actor_call section (the ones added with march_task_await are
  sufficient; the older duplicates are now redundant)

+++ runtime/march_scheduler.c
@@ -753,9 +753,9 @@ void march_cancel_token_cancel(march_cancel_token *tok) {
     atomic_store_explicit(&tok->cancelled, 1, memory_order_release);
 }
 
-int march_cancel_token_is_cancelled(march_cancel_token *tok) {
+int64_t march_cancel_token_is_cancelled(march_cancel_token *tok) {
     if (!tok) return 0;
-    return atomic_load_explicit(&tok->cancelled, memory_order_acquire);
+    return (int64_t)atomic_load_explicit(&tok->cancelled, memory_order_acquire);
 }
 
 void march_cancel_token_ref(march_cancel_token *tok) {
